added deploy to nightly test on test-ubuntu14

+++ cmake_modules/create_binary_archive.cmake
@@ -2,13 +2,12 @@
 #
 # Run this script by: cmake -P create_binary_archive.cmake
 
-if(CTEST_BINARY_DIRECTORY) # we are probably in a nighlty test
-  set(ARCHIVE_NAME "${HOSTNAME}_${TEST_NAME}")
-  set(ARCHIVE_PATH "${NIGHTLY_ARCHIVES_DIR}/${HOSTNAME}_${TEST_NAME}.tgz")
-  message("CTEST_BINARY_DIRECTORY=${CTEST_BINARY_DIRECTORY}")
-else(CTEST_BINARY_DIRECTORY)
+if(NOT CTEST_BINARY_DIRECTORY)
   message("CTEST_BINARY_DIRECTORY not defined: set to current dir '${CMAKE_CURRENT_SOURCE_DIR}'")
   set(CTEST_BINARY_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
+endif(NOT CTEST_BINARY_DIRECTORY)
+
+if(NOT DIST_FAMILY)
   # determine distro
   exec_program("${CTEST_BINARY_DIRECTORY}/share/scripts/distro.sh -c" OUTPUT_VARIABLE DISTRO_OUT)
   file(WRITE "${CTEST_BINARY_DIRECTORY}/CMakeFiles/distro_out.cmake" "${DISTRO_OUT}")
@@ -19,9 +18,14 @@ else(CTEST_BINARY_DIRECTORY)
   if(MAJOR_REV STREQUAL "")# its probably ubuntu ...
     set(MAJOR_REV ${REV})
   endif(MAJOR_REV STREQUAL "")
+endif(NOT DIST_FAMILY)
+
+if(HOSTNAME AND TEST_NAME)
+  # the script is probably included by a deploy from a nightly test
+  set(ARCHIVE_NAME "${HOSTNAME}_${TEST_NAME}")
+else()
   set(ARCHIVE_NAME "CFS_${DIST_FAMILY}_${MAJOR_REV}_${ARCH}")
-  set(ARCHIVE_PATH "${CTEST_BINARY_DIRECTORY}/${ARCHIVE_NAME}.tgz")
-endif(CTEST_BINARY_DIRECTORY)
+endif(HOSTNAME AND TEST_NAME)
 
 set(CFS_BIN_DIR "${CTEST_BINARY_DIRECTORY}/bin/${DIST_FAMILY}_${MAJOR_REV}_${ARCH}")
 message("  Searching for cfs executables in ${CFS_BIN_DIR}")
@@ -35,16 +39,18 @@ if(EXISTS "${CFS_BINARY}")
     "${CTEST_BINARY_DIRECTORY}/share"
   )
   # copy to subfolder
-  message("  Copy to subfolder '${ARCHIVE_NAME}' ...")
-  file(COPY ${FILES_TO_KEEP} DESTINATION ${ARCHIVE_NAME})
+  message("  Copy to subfolder '${CTEST_BINARY_DIRECTORY}/${ARCHIVE_NAME}' ...")
+  file(COPY ${FILES_TO_KEEP} DESTINATION "${CTEST_BINARY_DIRECTORY}/${ARCHIVE_NAME}")
   # find static archives and remove them
   file(GLOB_RECURSE STATIC_ARCHIVES "${CTEST_BINARY_DIRECTORY}/${ARCHIVE_NAME}/lib64/*.a")
-  file(REMOVE ${STATIC_ARCHIVES})  
+  if(STATIC_ARCHIVES)
+    file(REMOVE ${STATIC_ARCHIVES})
+  endif(STATIC_ARCHIVES)
 
   # pack the binaries
-  message("  Packing binaries to '${ARCHIVE_PATH}' ... ")
+  message("  Packing binaries to '${ARCHIVE_NAME}.tgz' ... ")
   EXECUTE_PROCESS(
-    COMMAND ${CMAKE_COMMAND} -E tar czf ${ARCHIVE_PATH} ${ARCHIVE_NAME}
+    COMMAND ${CMAKE_COMMAND} -E tar czf "${ARCHIVE_NAME}.tgz" ${ARCHIVE_NAME}
     WORKING_DIRECTORY "${CTEST_BINARY_DIRECTORY}"
   )
